+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Cover more functional tests on back end Web Service page
 - Cover more functional tests on back end Roles page
 - Cover more functional tests on back end Users page
+- Cover more functional tests on back end Groups page
 
 ## DEPENDENCIES